teamd-ttdp: add partial topocount support
Add support to supply a file which holds a list of possible topocounts. The
file shall contain one topocount value on each row and stored as hexadecimal,
i.e.:
   11223344
   aabbccdd
The topocount values in this list contains valid topocounts for which will
be used in recovery mode to let a ETBN join the backbone again if it
previously was lost.

Signed-off-by: Jonas Johansson <jonas.johansson@westermo.se>