libteam: Add patch from upstream master
Add patch from libteam upstream master that improves state transitions from
default to current. The reason only this patch is added for now is that it seem
to help with lacp issues we have seen on virtual systems using virtio.

The issue that has been observed is that an aggregate can be locked in a strange
state where one side of the lag is detached and the other distributing. This
also has effects on for instance FRNT and lag together. The underlying problem
with LACP LAGs and virtio might be more extensive but this patch at least seem
to stabilize the issue we have seen when rebooting a dut in a LAG FRNT ring
causes the ring and LAG to be permanently broken, because it cannot fully
negotiate at startup. Seemingly the LAG enters a state where it will not
correctly handle traffic. In a real system (or virtual with other link
emulation) this can still recover, but on virtio it seems that it will not.

This patch should be removed when we upgrade libteam in the future to 1.29.

The upstream patch on libteam is cc3d79f6c665eb7a9c14d9a3a453aee4ad23da4c:
--------------------------------------------------------------------------

teamd: lacp: send LACPDU when port state transitions from DEFAULT to CURRENT

Currently, when local port state become DEFAULTED, its Partner_State
field is set to 0x00

Bit 1 in Actor_State and Partner_State field is for LACP_Timeout. When
set, it means LACP fast is used.

If teamd_runner_lacp receives an LACPDU with Actor_State field bit 1
set, it transitions local port to CURRENT, and transmits LACPDU

If teamd_runner_lacp receives an LACPDU without Actor_State field bit 1
set, it transitions local port to CURRENT, but does not transmit LACPDU

The second scenario results in teamd starting data transmission when
partner port remains in either Defaulted or Expired state, causing
packets to be dropped by partner

Fix this by setting Partner_State field to 0x02 on the local port
when local port state transitions to DEFAULTED