teamdagentd: expect new path and name
Changing the expected names of confiuration files and pid
files from teamd- to lag-. The reason for this is because the
interfaces are in fact labelled lag- and this was a reason lag-mib
was broken due to the previous fixes. Also changed where teamdagentd
is expecting to find its pid-files from /var/run/teamd to /var/run.