teamd-ttdp: reset foreign lengthening
If a node has lengthening set from previously, and no longer has a neighbor,
it should clear the local lengthening flag in all cases. Previosuly, this
only happened if the node initially had a null neighbor; this change fixes
the behavior so that the flag is also correctly cleared for nodes that ended
up in a combined shortening+lengthening state.

Signed-off-by: Andy Koszela <andy.koszela@westermo.com>