teamd-ttdp: Always consider TOPOLOGY reception when determining end node status
Previously, we could sometimes not consider ourselves an end node,
even when we ought to, if we were quick enough to lose our neighbor
before the higher layers of the stack could notify us that we're no
longer receiving TOPOLOGY frames.

After this fix, we consider ourselves as an intermediate node only if
we both have a non-null neighbor (i.e. HELLO link status is up) and
we're recieving TOPOLOGY frames, as notified by the higher stack.

Signed-off-by: Andy Koszela <andy.koszela@westermo.com>